pyproject.toml
[project]
name = "bio-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.76.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.26.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.14",
]

[project.scripts]
build-cache = "bio_mcp.cache.build_cache:build_cache"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "I",   # isort (imports)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
